PR splits the current utils.py into logically-named files under src/dreamlayer/utils/, updates every import in repo via ruff --fix-import-sort, and adds a one-paragraph “module map” to docs/dev_guide.md. CI passes with no broken imports.
